user 테이블 생성_2
