+++ students/hyein/users/tests.py
@@ -1,12 +1,3 @@
-from django.db import models
+from django.test import TestCase
 
-class User(models.Model): 
-    name         = models.CharField(max_length=20)
-    email        = models.EmailField(max_length=200,unique=True)
-    password     = models.CharField(max_length=50)
-    phone_number = models.CharField(max_length=50)
-    created_at   = models.DateTimeField(auto_now_add=True)
-    updated_at   = models.DateTimeField(auto_now=True)
-    
-    class Meta: 
-        db_table = 'users'
+# Create your tests here.
